Rename `MelvinTarget` to `MelvinTargetable` to match other components
